Massive Streamlining

+++ commands/Chat/spoiler.js
@@ -1,3 +1,6 @@
+const Discord = require('discord.js');
+const fs = require('fs');
+
 module.exports = {
     name: 'spoiler',
     alias: ['spoiler', 'hide', 'redacted'],
@@ -11,18 +14,30 @@ module.exports = {
         '',
         'This command is effective if the input does not have any spoiler tags.'
     ],
-    run: async (msg, args) => {
+    dev: false,
+    mod: false,
+    activate: true,
+    /**
+     * @param {Discord.Message} msg The Discord.Message() object.
+     * @param {Array<String>} [args] The argument.
+     * @param {Map<String,any> | Discord.Collection<String|any>} [col] The collector.
+     */
+    run: async (msg, args, col) => {
+        const svr = JSON.parse(fs.readFileSync(`./data/guilds/${msg.guild.id}.json`));
+        if(!svr.modules.includes(module.exports.name)) return msg.channel.send('This module is not activated. Please activate it via the `setup` command.');
+
         if (args.length > 1) {
             let Mode = args.pop();
             let simple = ['s', 'simple'];
             let full = ['f', 'full'];
+            let word;
 
             if (simple.includes(Mode.toLowerCase())) {
-                var word = args.join('|| ||');
+                word = args.join('|| ||');
             } else if (full.includes(Mode.toLowerCase())) {
-                var word = [...`${args}`].join('').split('').join('||||').split(',').join(' ');
+                word = [...args.join(' ')].join('||||');
             } else {
-                var word = `${args.join('|| ||')}|| ||${Mode}`;
+                word = `${args.join('|| ||')}|| ||${Mode}`;
             }
 
             msg.delete().catch(e => console.log(e));
@@ -32,7 +47,7 @@ module.exports = {
             msg.delete().catch(e => console.log(e));
             return msg.channel.send(`||${word}||`);
         } else {
-            return msg.channel.send(`Not much stuff to spoil apparently...`);
+            return msg.channel.send(`Please provide a word to create a spoiler with.`);
         }
     }
 }
